Levels are now of type double
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gudhi/branches/cc_wit_hiraoka@3150 636b058d-ea47-450e-bf9e-a15bfbe3eedb

Former-commit-id: 996150ea3bdc6adac0d4f55a01bc10929f3a0abf

+++ src/Coxeter_triangulation/example/coxeter_complex.cpp
@@ -50,9 +50,9 @@ int main(int argc, char * const argv[]) {
     return 0;
   }
   Point_vector point_vector;
-  int init_level = 1;
+  double init_level = 1;
   if (argc == 3)
-    init_level = atoi(argv[2]);
+    init_level = atof(argv[2]);
   // Gudhi::Points_off_reader<Point_d> off_reader(argv[1]);
   // if (!off_reader.is_valid()) {
   //     std::cerr << "Coxeter triangulations - Unable to read file " << argv[1] << "\n";
